third_party/blink/web_tests/fast/forms/number/number-l10n-input-expected.txt

Putting some ASCII digit strings to a number input using localized digits.
PASS document.execCommand("InsertText", false, "1234"); arabicInput.value is "1234"
PASS document.execCommand("InsertText", false, "1.234"); arabicInput.value is "1.234"
PASS document.execCommand("InsertText", false, "1,234"); arabicInput.value is "1234"


Putting some digit strings to a French number input.
PASS document.execCommand("InsertText", false, "1234"); frenchInput.value is "1234"
PASS document.execCommand("InsertText", false, "1,234"); frenchInput.value is "1.234"
PASS document.execCommand("InsertText", false, "1.234"); frenchInput.value is "1.234"


Puttting some digit strings to an English number input.
PASS document.execCommand("InsertText", false, "1234"); englishInput.value is "1234"
PASS document.execCommand("InsertText", false, "1.234"); englishInput.value is "1.234"
PASS document.execCommand("InsertText", false, "1,234"); englishInput.value is "1234"
PASS successfullyParsed is true

TEST COMPLETE

  
